+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: RAG Builder object
 - Change: The Prompt Builder object now filters for projects with the Prompt-Engineering tag
 - Change: The Prompt Builder object now supports deprecation of LLMs
+- Fix: Instead of using hyphens, invalid characters are removed fully for *validate-python-package-name.js*
 
 ## SAS Portal Framework for SAS Viya v1.2.2